Update Goodies to use new Dates Role format
Functionality hasn't changed except to keep in line with the changes to
the Dates GoodieRole - i.e., correcting mistakes.

I would like to get a bunch of the tests removed from many of these
Goodies that are *just* testing functionality of the Role (@mintsoft -
you've mentioned this before I believe). We should focus on *just*
testing the Goodie, and letting the maintainers know if any issues arise
rather than adding more tests that don't do anything but use a different
date format.

Also, we should really move away from the DD/MM/YYYY and MM/DD/YYYY
format in tests - it is a pain as we have to set language/location if we
want it to play nicely; instead we should be using more descriptive
dates that would be representative of the real world (Jan 1st 2012
etc.).